refactor: update hocs index file
